+++ content-types/content-type-read-receipt/src/ReadReceipt.test.ts
@@ -1,10 +1,37 @@
-import { Client } from "@xmtp/xmtp-js";
-import { Wallet } from "ethers";
+import { getRandomValues } from "node:crypto";
+import { Client, IdentifierKind, type Signer } from "@xmtp/node-sdk";
+import { createWalletClient, http, toBytes } from "viem";
+import { generatePrivateKey, privateKeyToAccount } from "viem/accounts";
+import { sepolia } from "viem/chains";
 import {
   ContentTypeReadReceipt,
   ReadReceiptCodec,
   type ReadReceipt,
 } from "./ReadReceipt";
+
+const testEncryptionKey = getRandomValues(new Uint8Array(32));
+
+export const createSigner = (): Signer => {
+  const account = privateKeyToAccount(generatePrivateKey());
+  const wallet = createWalletClient({
+    account,
+    chain: sepolia,
+    transport: http(),
+  });
+  return {
+    type: "EOA",
+    getIdentifier: () => ({
+      identifierKind: IdentifierKind.Ethereum,
+      identifier: account.address.toLowerCase(),
+    }),
+    signMessage: async (message: string) => {
+      const signature = await wallet.signMessage({
+        message,
+      });
+      return toBytes(signature);
+    },
+  };
+};
 
 describe("ReadReceiptContentType", () => {
   it("has the right content type", () => {
@@ -15,40 +42,35 @@ describe("ReadReceiptContentType", () => {
   });
 
   it("can send a read receipt", async () => {
-    const aliceWallet = Wallet.createRandom();
-    const aliceClient = await Client.create(aliceWallet, {
+    const signer1 = createSigner();
+    const client1 = await Client.create(signer1, testEncryptionKey, {
       codecs: [new ReadReceiptCodec()],
       env: "local",
     });
-    await aliceClient.publishUserContact();
 
-    const bobWallet = Wallet.createRandom();
-    const bobClient = await Client.create(bobWallet, {
+    const signer2 = createSigner();
+    const client2 = await Client.create(signer2, testEncryptionKey, {
       codecs: [new ReadReceiptCodec()],
       env: "local",
     });
-    await bobClient.publishUserContact();
 
-    const conversation = await aliceClient.conversations.newConversation(
-      bobWallet.address,
-    );
+    const dm = await client1.conversations.newDm(client2.inboxId);
 
     const readReceipt: ReadReceipt = {};
 
-    await conversation.send(readReceipt, {
-      contentType: ContentTypeReadReceipt,
-    });
+    await dm.send(readReceipt, ContentTypeReadReceipt);
 
-    const bobConversation = await bobClient.conversations.newConversation(
-      aliceWallet.address,
-    );
-    const messages = await bobConversation.messages();
+    await client2.conversations.sync();
+    const dms = client2.conversations.listDms();
 
+    expect(dms.length).toBe(1);
+
+    await dms[0].sync();
+    const messages = await dms[0].messages();
     expect(messages.length).toBe(1);
 
     const readReceiptMessage = messages[0];
-    const messageContent = readReceiptMessage.contentType;
-    expect(messageContent.typeId).toBe("readReceipt");
+    expect(readReceiptMessage.contentType?.typeId).toBe("readReceipt");
   });
 
   it("has a proper shouldPush value", () => {