+++ packages/frames-validator/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./validator.js";
-export * from "./validation.js";