+++ sdks/js-sdk/src/utils/typedefs.ts
@@ -1,5 +1,0 @@
-export type Flatten<T> = {
-  [K in keyof T]: T[K];
-};
-
-export type WithoutUndefined<T> = { [P in keyof T]: NonNullable<T[P]> };