+++ content-types/content-type-reply/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @xmtp/content-type-reply
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [340fcf4]
+  - @xmtp/content-type-primitives@2.0.1
+  - @xmtp/proto@3.78.0
+
 ## 2.0.0
 
 ### Major Changes