+++ sdks/browser-sdk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @xmtp/browser-sdk
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [340fcf4]
+  - @xmtp/content-type-group-updated@2.0.1
+  - @xmtp/content-type-primitives@2.0.1
+  - @xmtp/content-type-text@2.0.1
+  - @xmtp/proto@3.78.0
+  - uuid@11.1.0
+
 ## 1.1.0
 
 ### Minor Changes